requirements.txt
numpy==1.24.3
h5py==3.11.0
matplotlib==3.7.2
pandas==2.0.2
tqdm==4.65.0
scipy==1.13.1
torchvision==0.15.2
scikit-learn==1.5.0
torchviz==0.0.2